+++ src/components/Desktop/clock/clock.module.scss
@@ -1,0 +1,10 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  
+  font-size: .9rem;
+  font-weight: 400;
+  color: white;
+}
