--- src/components/clock/clock.module.scss
@@ -1,10 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  
-  font-size: .9rem;
-  font-weight: 500;
-  color: white;
-}
